services/replicate-poc/pyproject.toml
[tool.poetry]
name = "replicate-poc"
version = "0.1.0"
description = "PostgreSQL replication POC service"
authors = ["Jeffrey Wescott <jeffrey.wescott@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.13"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 